// src/api/models/sys-user-input.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 全部
 * 全部API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { AdminType } from './admin-type';
import { CommonStatus } from './common-status';
import { Gender } from './gender';

/**
 * 用户输出表
 * @export
 * @interface SysUserInput
 */
export interface SysUserInput {
    /**
     * 
     * @type {number}
     * @memberof SysUserInput
     */
    ID?: number;
    /**
     * 账号
     * @type {string}
     * @memberof SysUserInput
     */
    Account?: string | null;
    /**
     * 工号
     * @type {string}
     * @memberof SysUserInput
     */
    Code?: string | null;
    /**
     * FirstName
     * @type {string}
     * @memberof SysUserInput
     */
    FirstName?: string | null;
    /**
     * LastName
     * @type {string}
     * @memberof SysUserInput
     */
    LastName?: string | null;
    /**
     * 姓名
     * @type {string}
     * @memberof SysUserInput
     */
    Name?: string | null;
    /**
     * 
     * @type {Gender}
     * @memberof SysUserInput
     */
    Sex?: Gender;
    /**
     * 邮箱
     * @type {string}
     * @memberof SysUserInput
     */
    Email?: string | null;
    /**
     * 手机
     * @type {string}
     * @memberof SysUserInput
     */
    Phone?: string | null;
    /**
     * 
     * @type {AdminType}
     * @memberof SysUserInput
     */
    AdminType?: AdminType;
    /**
     * 机构ID
     * @type {number}
     * @memberof SysUserInput
     */
    OrgID?: number | null;
    /**
     * 
     * @type {CommonStatus}
     * @memberof SysUserInput
     */
    Status?: CommonStatus;
}