// src/api/models/sys-user-output.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 全部
 * 全部API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { AdminType } from './admin-type';
import { CommonStatus } from './common-status';
import { Gender } from './gender';

/**
 * 用户输出表
 * @export
 * @interface SysUserOutput
 */
export interface SysUserOutput {
    /**
     * ID
     * @type {number}
     * @memberof SysUserOutput
     */
    ID?: number;
    /**
     * 备用字段1
     * @type {string}
     * @memberof SysUserOutput
     */
    Reserved1?: string | null;
    /**
     * 更新时间
     * @type {string}
     * @memberof SysUserOutput
     */
    UpdatedTime?: string | null;
    /**
     * 修改者Id
     * @type {string}
     * @memberof SysUserOutput
     */
    UpdatorCode?: string | null;
    /**
     * 修改者名称
     * @type {string}
     * @memberof SysUserOutput
     */
    UpdatorName?: string | null;
    /**
     * 创建时间
     * @type {string}
     * @memberof SysUserOutput
     */
    CreatedTime?: string | null;
    /**
     * 创建者Id
     * @type {string}
     * @memberof SysUserOutput
     */
    CreatorCode?: string | null;
    /**
     * 创建者名称
     * @type {string}
     * @memberof SysUserOutput
     */
    CreatorName?: string | null;
    /**
     * 账号
     * @type {string}
     * @memberof SysUserOutput
     */
    Account?: string | null;
    /**
     * 工号
     * @type {string}
     * @memberof SysUserOutput
     */
    Code?: string | null;
    /**
     * FirstName
     * @type {string}
     * @memberof SysUserOutput
     */
    FirstName?: string | null;
    /**
     * LastName
     * @type {string}
     * @memberof SysUserOutput
     */
    LastName?: string | null;
    /**
     * 姓名
     * @type {string}
     * @memberof SysUserOutput
     */
    Name?: string | null;
    /**
     * 
     * @type {Gender}
     * @memberof SysUserOutput
     */
    Sex?: Gender;
    /**
     * 邮箱
     * @type {string}
     * @memberof SysUserOutput
     */
    Email?: string | null;
    /**
     * 手机
     * @type {string}
     * @memberof SysUserOutput
     */
    Phone?: string | null;
    /**
     * 
     * @type {AdminType}
     * @memberof SysUserOutput
     */
    AdminType?: AdminType;
    /**
     * 机构ID
     * @type {number}
     * @memberof SysUserOutput
     */
    OrgID?: number | null;
    /**
     * 机构名称
     * @type {string}
     * @memberof SysUserOutput
     */
    OrgName?: string | null;
    /**
     * 角色名称
     * @type {string}
     * @memberof SysUserOutput
     */
    RoleNames?: string | null;
    /**
     * 
     * @type {CommonStatus}
     * @memberof SysUserOutput
     */
    Status?: CommonStatus;
}